requirements.txt
blinker==1.4
-e git://github.com/jarus/flask-testing/@ef069cab65bbbbd866a83b7b3b3aa4e57d970294#egg=flask-testing
pyopenssl==20.0.1
coverage==5.4
pre-commit==2.10.1
pydealer==1.4.0
matplotlib==3.3.4
pandas==1.2.2
dateparser==1.0.0
discord.py==1.6.0
scipy==1.6.0
bleach==3.3.0
certifi==2020.12.5
chardet==3.0.4
Click==7.1.2
dnspython==1.16.0
eventlet==0.30.1
Flask==1.1.2
Flask-SocketIO==5.0.1
github-webhook==1.0.4
greenlet==1.0.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
cryptography
monotonic==1.5
numexpr==2.7.2
numpy==1.20.1
python-engineio==4.0.0
python-socketio==5.0.4
requests==2.25.1
six==1.15.0
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
setuptools==53.0.0
ply==3.11
pytz==2021.1
